requirements.txt
bs4==0.0.2
get_project_root==0.2
lucide-fasthtml==0.0.9
python-dotenv==1.0.1
python-fasthtml==0.9.0
requests==2.32.3
supabase==2.10.0
